z3: use separate expression types for different sorts
This changes the Expr type to be an interface (with a small method
set) and introduces separate concrete types for each kind (in the Sort
sense) of expression.

This lets us separate the methods that apply to different kinds into
different types, rather than mashing them all together on Expr. This
allows for nicer method names. It declutters the documentation and
also simplifies it since many of the previously implicit constraints
like "l must have boolean sort" and now explicit in the type
signature. And it adds some level of static type checking. The only
obvious downside is that some generic operations like IfThenElse now
require an explicit type assertion on the result, since we can't
express even their kinding in the Go type system.